[#18] Refactor : Diary와 DiaryList 도메인 통합 - DiaryLIstController -> DiaryController - DiaryListService -> DiaryService - DiaryController의 mapping url : /diary -> /diaries
