+++ src/main/java/mashup/backend/etc/our/diary/controller/DiaryController.java
@@ -1,20 +1,27 @@
 package mashup.backend.etc.our.diary.controller;
 
-import mashup.backend.etc.our.diary.dto.ReqDeleteDiaryDto;
-import mashup.backend.etc.our.diary.dto.ReqPostDiaryDto;
-import mashup.backend.etc.our.diary.dto.ResDiaryDto;
+import lombok.RequiredArgsConstructor;
+import mashup.backend.etc.our.diary.dto.*;
 import mashup.backend.etc.our.diary.service.DiaryService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+@RequiredArgsConstructor
 @RestController
-@RequestMapping("/diary")
+@RequestMapping("/diaries")
 public class DiaryController {
-    private DiaryService diaryService;
 
-    public DiaryController(DiaryService diaryService){
-        this.diaryService = diaryService;
+    private final DiaryService diaryService;
+
+    @GetMapping("/")
+    public ResDiaryListDto readDiaryList(@RequestBody ReqDiaryListDto reqDiaryListDto){
+        return diaryService.readDiaryList(reqDiaryListDto);
+    }
+
+    @GetMapping("/{diaryId}")
+    public ResSingleDiaryDto getSingleDiaryById(@PathVariable Long diaryId, @RequestBody ReqSingleDiaryDto reqSingleDiaryDto){
+        return diaryService.getSingleDiaryById(diaryId,reqSingleDiaryDto);
     }
 
     @PostMapping
